fix 64bit warnings

+++ include/crow/dumb_timer_queue.h
@@ -16,7 +16,7 @@ namespace crow
         class dumb_timer_queue
         {
         public:
-            using key = std::pair<dumb_timer_queue*, int>;
+            using key = std::pair<dumb_timer_queue*, intptr_t>;
 
             void cancel(key& k)
             {
@@ -33,7 +33,7 @@ namespace crow
             key add(std::function<void()> f)
             {
                 dq_.emplace_back(std::chrono::steady_clock::now(), std::move(f));
-                int ret = step_+dq_.size()-1;
+                intptr_t ret = step_+dq_.size()-1;
 
                 CROW_LOG_DEBUG << "timer add inside: " << this << ' ' << ret ;
                 return {this, ret};
